test get data

+++ lib/movies/data/model/movies_model.dart
@@ -17,7 +17,8 @@ class MoviesModel extends Movie {
         backdropPath: json['backdrop_path'],
         genreIds: List<int>.from(json['genre_ids'].map((e) => e)),
         overView: json['overview'],
-        voteAverage: json['vote_average'],
+        //TODO : CHICK THIS
+        voteAverage: json['vote_average'].toDouble(),
         releaseDate: json['release_date'],
       );
 }
